pyproject.toml
[project]
name = "autouploadbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiogram>=3.24.0",
    "fastapi>=0.128.0",
    "httpx>=0.28.1",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.40.0",
    "yt-dlp>=2025.12.8",
]
